:recycle: Fix hardware wallet account initialisation check
